Rename and move VariableListenersWithSources
