+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/custom/PiggybackShadowVariableDescriptor.java
@@ -13,6 +13,7 @@ import org.optaplanner.core.impl.domain.entity.descriptor.EntityDescriptor;
 import org.optaplanner.core.impl.domain.policy.DescriptorPolicy;
 import org.optaplanner.core.impl.domain.variable.descriptor.ShadowVariableDescriptor;
 import org.optaplanner.core.impl.domain.variable.descriptor.VariableDescriptor;
+import org.optaplanner.core.impl.domain.variable.listener.VariableListenerWithSources;
 import org.optaplanner.core.impl.domain.variable.supply.Demand;
 import org.optaplanner.core.impl.score.director.InnerScoreDirector;
 
@@ -101,7 +102,8 @@ public class PiggybackShadowVariableDescriptor<Solution_> extends ShadowVariable
     }
 
     @Override
-    public Iterable<ListenerSources<Solution_>> buildVariableListener(InnerScoreDirector<Solution_, ?> scoreDirector) {
+    public Iterable<VariableListenerWithSources<Solution_>>
+            buildVariableListeners(InnerScoreDirector<Solution_, ?> scoreDirector) {
         throw new UnsupportedOperationException("The piggybackShadowVariableDescriptor (" + this
                 + ") cannot build a variable listener.");
     }
