+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/listener/VariableListenerWithSources.java
@@ -1,4 +1,4 @@
-package org.optaplanner.core.impl.domain.variable.custom;
+package org.optaplanner.core.impl.domain.variable.listener;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -6,18 +6,23 @@ import java.util.Collections;
 import org.optaplanner.core.api.domain.variable.AbstractVariableListener;
 import org.optaplanner.core.impl.domain.variable.descriptor.VariableDescriptor;
 
-public final class ListenerSources<Solution_> {
+/**
+ * Holds a variable listener and all its source variable descriptors.
+ *
+ * @param <Solution_>
+ */
+public final class VariableListenerWithSources<Solution_> {
 
     private final AbstractVariableListener<Solution_, Object> variableListener;
     private final Collection<VariableDescriptor<Solution_>> sourceVariableDescriptors;
 
-    public ListenerSources(AbstractVariableListener<Solution_, Object> variableListener,
+    public VariableListenerWithSources(AbstractVariableListener<Solution_, Object> variableListener,
             Collection<VariableDescriptor<Solution_>> sourceVariableDescriptors) {
         this.variableListener = variableListener;
         this.sourceVariableDescriptors = sourceVariableDescriptors;
     }
 
-    public ListenerSources(AbstractVariableListener<Solution_, Object> variableListener,
+    public VariableListenerWithSources(AbstractVariableListener<Solution_, Object> variableListener,
             VariableDescriptor<Solution_> sourceVariableDescriptor) {
         this(variableListener, Collections.singleton(sourceVariableDescriptor));
     }
@@ -30,7 +35,7 @@ public final class ListenerSources<Solution_> {
         return sourceVariableDescriptors;
     }
 
-    public Collection<ListenerSources<Solution_>> toCollection() {
+    public Collection<VariableListenerWithSources<Solution_>> toCollection() {
         return Collections.singleton(this);
     }
 }
