+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/nextprev/PreviousElementShadowVariableDescriptor.java
@@ -7,7 +7,7 @@ import org.optaplanner.core.api.domain.variable.AbstractVariableListener;
 import org.optaplanner.core.api.domain.variable.PreviousElementShadowVariable;
 import org.optaplanner.core.impl.domain.common.accessor.MemberAccessor;
 import org.optaplanner.core.impl.domain.entity.descriptor.EntityDescriptor;
-import org.optaplanner.core.impl.domain.variable.custom.ListenerSources;
+import org.optaplanner.core.impl.domain.variable.listener.VariableListenerWithSources;
 import org.optaplanner.core.impl.score.director.InnerScoreDirector;
 
 public class PreviousElementShadowVariableDescriptor<Solution_>
@@ -35,8 +35,9 @@ public class PreviousElementShadowVariableDescriptor<Solution_>
     }
 
     @Override
-    public Iterable<ListenerSources<Solution_>> buildVariableListener(InnerScoreDirector<Solution_, ?> scoreDirector) {
-        return new ListenerSources<>(new PreviousElementVariableListener<>(this, sourceVariableDescriptor),
+    public Iterable<VariableListenerWithSources<Solution_>>
+            buildVariableListeners(InnerScoreDirector<Solution_, ?> scoreDirector) {
+        return new VariableListenerWithSources<>(new PreviousElementVariableListener<>(this, sourceVariableDescriptor),
                 sourceVariableDescriptor).toCollection();
     }
 }
